add augparam to config and so on
