+++ ex_9/kajiwara/src/dataset.py
@@ -5,17 +5,12 @@ import pandas as pd
 import librosa
 from torch.utils.data import Dataset
 
+from omegaconf import DictConfig
+
 import augmentations
 
 
 def mel_spec(input: np.ndarray, sr: int, win_size: int, hop_len: int, n_mels: int, ) -> np.ndarray:
-    # spec = librosa.stft(
-    #     y=input, n_fft=win_size, win_length=win_size, hop_length=hop_len)
-    # spec = np.abs(spec) ** 2.0
-
-    # mel_filter_bank = librosa.filters.mel(sr=sr, n_fft=win_size, n_mels=n_mels)
-    # mel = np.dot(mel_filter_bank, spec)
-
     mel = librosa.feature.melspectrogram(
         y=input, sr=sr, n_mels=n_mels, n_fft=win_size, win_length=win_size, hop_length=hop_len)
     log_mel = librosa.amplitude_to_db(mel)
@@ -26,15 +21,13 @@ def mel_spec(input: np.ndarray, sr: int, win_size: int, hop_len: int, n_mels: in
 class FSDDataset(Dataset):
     def __init__(
         self, audio_path: str, metadata_path: str,
-        win_size_rate: float, overlap: float, n_mels: int, training: bool = True,
-        gaussian_noise_snr: bool = True,
-        time_shift: bool = True,
-        volume_control: bool = True
+        win_size_rate: float, overlap: float, n_mels: int,
+        aug_cfg: DictConfig, training: bool = True,
     ):
         self.training = training
-        self.gaussian_noise_snr = gaussian_noise_snr
-        self.time_shift = time_shift
-        self.volume_control = volume_control
+        self.gns_cfg = aug_cfg['gaussian_noise']
+        self.ts_cfg = aug_cfg['time_shift']
+        self.vc_cfg = aug_cfg['volume_control']
 
         self.audio_path = Path(audio_path)
 
@@ -60,14 +53,14 @@ class FSDDataset(Dataset):
         else:
             waveform = waveform[:int(1.0*sr)]
 
-        if self.gaussian_noise_snr:
-            augmentations.gaussian_noise_snr(waveform)
+        if self.gns_cfg['using']:
+            augmentations.gaussian_noise_snr(waveform, self.gns_cfg['min_snr'], self.gns_cfg['max_snr'])
 
-        if self.time_shift:
-            augmentations.time_shift(waveform, sr)
+        if self.ts_cfg['using']:
+            augmentations.time_shift(waveform, sr, self.ts_cfg['max_shift_sec'], self.ts_cfg['padding_mode'])
 
-        if self.volume_control:
-            augmentations.volume_control(waveform)
+        if self.vc_cfg['using']:
+            augmentations.volume_control(waveform, self.vc_cfg['db_lim'], self.vc_cfg['mode'])
 
         win_size = int(self.win_size_rate * sr)
         feature = mel_spec(waveform, sr, win_size,
