+++ ex_9/kajiwara/src/run.py
@@ -14,84 +14,17 @@ from torch.utils.tensorboard import SummaryWriter
 
 from dataset import FSDDataset, EXFSD_Dataset
 from model import ConformerModel, GRUModel
+from traininig import train, valid
 
 TIME_TEMPLATE = '%Y%m%d%H%M%S'
 
 
-def train(trainloader, optimizer, device, global_step,  model, criterion, writer, fold):
-    model.train()
-
-    n_batch = len(trainloader)
-    train_loss = 0
-    train_acc = 0
-    total = 0
-    for batch_num, (t_data, labels) in enumerate(trainloader):
-        optimizer.zero_grad()
-        t_data = t_data.to(device)
-        labels = labels.to(device)
-
-        outputs = model(t_data)
-
-        optimizer.zero_grad()
-
-        loss = criterion(outputs, labels)
-        loss.backward()
-        optimizer.step()
-        train_loss += loss.item()
-
-        _, predict = torch.max(outputs.data, 1)
-        total += labels.size(0)
-        correct = (predict == labels).sum()
-        train_acc += correct.item()
-
-        writer.add_scalar(f"{fold}/loss", loss.item(), global_step)
-        writer.add_scalar(f"{fold}/acc", correct.item() /
-                          labels.size(0), global_step)
-        print(
-            f'batch: {batch_num}/{n_batch}, '
-            f'loss: {loss.item()}, train loss: {train_loss/(batch_num+1)}, '
-            f'acc: {correct.item()/labels.size(0)}, train acc: {train_acc/total} ')
-        global_step += 1
-
-    train_loss /= n_batch
-    train_acc /= len(trainloader.dataset)
-
-    return global_step, train_loss, train_acc
-
-
-def valid(validloader, device, model, criterion):
-    model.eval()
-
-    valid_loss = 0
-    valid_acc = 0
-    total = 0
-
-    with torch.no_grad():
-        for i, (t_data, labels) in enumerate(validloader):
-            t_data = t_data.to(device)
-            labels = labels.to(device)
-
-            outputs = model(t_data)
-
-            loss = criterion(outputs, labels)
-            valid_loss += loss.item()
-
-            _, predict = torch.max(outputs.data, 1)
-            total += labels.size(0)
-            correct = (predict == labels).sum()
-            valid_acc += correct.item()
-
-        valid_loss /= len(validloader)
-        valid_acc /= total
-
-    print(f'val loss: {valid_loss}, val acc: {valid_acc}')
-
-
-@ hydra.main(config_path='../config', config_name='train')
+@ hydra.main(config_path='../config', config_name='param')
 def run(cfg):
     """set config"""
     path_cfg = cfg['path']
     audio_cfg = cfg['audio']
+    aug_cfg = cfg['augmentation']
     train_cfg = cfg['training']
 
     """set path"""
@@ -165,7 +98,19 @@ def run(cfg):
             win_size_rate=win_size_rate,
             overlap=overlap,
             n_mels=n_mels,
+            aug_cfg=aug_cfg
         )
+
+    testset = FSDDataset(
+        audio_path=audio_path,
+        metadata_path=test_metadata_path,
+        training=False,
+        win_size_rate=win_size_rate,
+        overlap=overlap,
+        n_mels=n_mels,
+        aug_cfg=aug_cfg
+    )
+
     idxes = [i for i in range(len(dataset))]
     kf = KFold(n_splits=kfold, shuffle=True)
     preds_by_fold = []
@@ -212,16 +157,7 @@ def run(cfg):
 
         """prediction"""
         model.eval()
-        testset = FSDDataset(
-            audio_path=audio_path,
-            metadata_path=test_metadata_path,
-            training=False,
-            win_size_rate=win_size_rate,
-            overlap=overlap,
-            n_mels=n_mels
-        )
-        testloader = DataLoader(
-            testset, batch_size=16, pin_memory=True)
+        testloader = DataLoader(testset, batch_size=16, pin_memory=True)
         preds = []
         for data in testloader:
             data = data.to(device)
